Rename app slice and move api functions
