+++ src/navigation/RootStackNavigation.tsx
@@ -46,7 +46,7 @@ import SwitchNetworkScreen from '../screens/SwitchNetworkScreen'
 import SwitchWalletAfterDeletionScreen from '../screens/SwitchWalletAfterDeletionScreen'
 import SwitchWalletScreen from '../screens/SwitchWalletScreen'
 import TransactionScreen from '../screens/TransactionScreen'
-import { routeChanged } from '../store/appMetadataSlice'
+import { routeChanged } from '../store/appSlice'
 import { isNavStateRestorable, rootStackNavigationRef } from '../utils/navigation'
 import InWalletTabsNavigation from './InWalletNavigation'
 import RootStackParamList from './rootStackRoutes'
