+++ src/store/addressesSlice.ts
@@ -34,8 +34,8 @@ import { AddressToken } from '../types/tokens'
 import { PendingTransaction } from '../types/transactions'
 import { getRandomLabelColor } from '../utils/colors'
 import { extractNewTransactions, extractRemainingPendingTransactions } from '../utils/transactions'
-import { appReset } from './actions'
 import { newWalletGenerated, walletUnlocked } from './activeWalletSlice'
+import { appReset } from './appSlice'
 import { customNetworkSettingsSaved, networkPresetSwitched } from './networkSlice'
 import { RootState } from './store'
